[wrangler] refactor: make types command definition more consistent
for consistency with (most of) the rest of the wrangler command
use a `typesOptions` and `typesHandler` arguments instead of
inlining such functions in the `wrangler.command()` call
